:hammer: probe the CLI symlink once at startup and move the entry to the extension page
