+++ app/src/desktopMain/kotlin/com/crosspaste/ui/extension/ExtensionContentView.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -15,6 +16,8 @@ import com.composables.icons.materialsymbols.MaterialSymbols
 import com.composables.icons.materialsymbols.rounded.Block
 import com.composables.icons.materialsymbols.rounded.Code
 import com.composables.icons.materialsymbols.rounded.Document_scanner
+import com.crosspaste.cli.CliSymlinkService
+import com.crosspaste.cli.CliSymlinkState
 import com.crosspaste.ui.LocalThemeExtState
 import com.crosspaste.ui.MCP
 import com.crosspaste.ui.NavigationManager
@@ -31,8 +34,13 @@ import org.koin.compose.koinInject
 
 @Composable
 fun ExtensionContentView() {
+    val cliSymlinkService = koinInject<CliSymlinkService>()
     val navigateManager = koinInject<NavigationManager>()
     val themeExt = LocalThemeExtState.current
+
+    // Probed once at app startup; NOT_SUPPORTED covers non-macOS platforms
+    // and dev runs without the bundled CLI binary
+    val cliSymlinkState by cliSymlinkService.state.collectAsState()
 
     var isProxyExpanded by remember { mutableStateOf(false) }
 
@@ -91,6 +99,19 @@ fun ExtensionContentView() {
                         navigateManager.navigate(SourceControl)
                     },
                 )
+            }
+        }
+
+        // macOS-only `crosspaste` terminal command install/repair (D6 PATH integration)
+        if (cliSymlinkState != CliSymlinkState.NOT_SUPPORTED) {
+            item {
+                SectionHeader("command_line", topPadding = medium)
+            }
+
+            item {
+                SettingSectionCard {
+                    CliToolSettingItem()
+                }
             }
         }
     }
